Don't choke on `map`, `splice`, and other array methods + make it reactive (#2)
* Move side effects to a separate method

Inside `constructor()`, `this` is bound to the original object, not the reactive Proxy object. All side effects should be moved outside `constructor()` to make the Vue reactivity system work.

See https://stackoverflow.com/a/67895251/5990035

* Fix buttons' hints

* Refactor to `async`/`await`

We don't need to await init() to resolve—Vue will do everything for us.

Moreover, if mounted() is async, it breaks the `<madata-auth>` component.